Readjusted namespaces, removed Geometry example as it's no longer relevant.
