+++ wgpu/src/triangle/gradient.rs
@@ -1,4 +1,4 @@
-use crate::buffers::dynamic_buffers::DynamicBuffer;
+use crate::buffers::dynamic::DynamicBuffer;
 use crate::settings;
 use crate::triangle::{
     default_fragment_target, default_multisample_state,
